+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Use of a list of attributes to map protein id and gene in the GFF.
+- Improve protein id and gene id mapping using a list of attributes. 
 
 ## [v0.1.0] – 2025-04-29